+++ accounts/api/serializers.py
@@ -1,6 +1,5 @@
-from xml.dom import ValidationErr
-
 from accounts.models import City, Profile, Project, Role, Service, User
+from accounts.utils import base64_file
 from django.contrib.auth import authenticate
 from rest_framework import serializers
 
@@ -90,6 +89,32 @@ class ProfileSerializer(serializers.ModelSerializer):
         fields = "__all__"
         read_only_fields = ["is_owner", "projects"]
 
+    def to_internal_value(self, data):
+        profile_picture = data.pop("profile_picture", None)
+        initial = super().to_internal_value(data)
+
+        if profile_picture is not None:
+            image = base64_file(profile_picture)
+            if image:
+                initial["profile_picture"] = image
+
+        return initial
+
+    def to_representation(self, instance):
+
+        initial = super().to_representation(instance)
+
+        profile_picture = initial.pop("profile_picture", None)
+
+        url = (
+            self.context["request"].build_absolute_uri(profile_picture)
+            if profile_picture
+            else None
+        )
+        initial["profile_picture"] = url
+        print(url)
+        return initial
+
     def get_is_owner(self, _):
         owner = self.context["is_owner"]
         return owner
@@ -101,10 +126,11 @@ class ProfileSerializer(serializers.ModelSerializer):
     def update(self, instance, validated_data):
         role = validated_data.get("role")
         services = validated_data.pop("services", [])
+
         instance = super().update(instance, validated_data)
+
         if role != None:
             instance.services.set(services, clear=True)
-        print(instance.services)
 
         return instance
