+++ frontend/src/components/registration/Personalize.js
@@ -11,15 +11,16 @@ import RadioGroup from "@mui/material/RadioGroup";
 import Select from "@mui/material/Select";
 import Stack from "@mui/material/Stack";
 import { styled } from "@mui/material/styles";
+import TextField from "@mui/material/TextField";
 import Typography from "@mui/material/Typography";
 import { updateProfile } from "api/auth";
-import { capitalCase } from "change-case";
 import useFormikValidation from "hooks/useFormikValidation";
-import React, { useState } from "react";
+import React, { useRef, useState } from "react";
 import * as yup from "yup";
+import { getUploadedFileValue } from "../../util/functions";
 
-const price_estimations = ["No", "Yes"];
-const cleanings = ["No", "Yes"];
+const price_estimations = [false, true];
+const cleanings = [false, true];
 const experiences = [1, 2, 3, 4, 5, 6, 7, 8];
 const team_sizes = ["1-3", "3+"];
 
@@ -31,6 +32,8 @@ const Personalize = function (props) {
   const [isSubmitting, setIsSubmitting] = useState(false);
   const [priceEstimationError, setPriceEstimationError] = useState("");
   const [cleaningError, setCleaningError] = useState("");
+  const ref = useRef();
+  const [image, setImage] = useState(null);
 
   const validationSchema = yup.object().shape({
     experience: yup
@@ -48,11 +51,11 @@ const Personalize = function (props) {
   const formik = useFormikValidation({
     initialValues: {
       experience: profile.experience || "",
-      price_estimate: profile.price_estimate || null,
-      cleaning: profile.cleaning || null,
-      team_size: "",
+      price_estimate: profile.price_estimate,
+      cleaning: profile.cleaning,
+      team_size: profile.team_size,
     },
-    onSubmit(values) {
+    onSubmit: async (values) => {
       let valid = true;
       if (values.price_estimate == null) {
         valid = false;
@@ -65,9 +68,10 @@ const Personalize = function (props) {
       }
       if (!valid) return;
       setIsSubmitting(true);
-      updateProfile(props.userId, values)
+      const profile_picture = image ? await getUploadedFileValue(image) : image;
+
+      updateProfile(props.userId, { ...values, profile_picture })
         .then(({ data }) => {
-          console.log(data);
           props.nextStep();
         })
         .catch(({ response }) => {})
@@ -82,6 +86,17 @@ const Personalize = function (props) {
 
   const handleSkip = () => {
     props.nextStep();
+  };
+
+  const handleImageChange = (event) => {
+    const files = Array.from(event.target.files);
+    const [file] = files;
+    const url = URL.createObjectURL(file);
+    setImage(url);
+  };
+
+  const handleRemoveImage = () => {
+    setImage(null);
   };
 
   const { getFieldProps, handleSubmit } = formik;
@@ -100,7 +115,10 @@ const Personalize = function (props) {
               alignItems: "center",
             }}>
             <Label sx={{ marginBottom: 1 }}>Avatar</Label>
-            <Avatar sx={{ width: 100, height: 100 }} />
+            <Avatar
+              sx={{ width: 100, height: 100 }}
+              src={image || profile.profile_picture}
+            />
           </Box>
           <Box
             sx={{
@@ -109,8 +127,16 @@ const Personalize = function (props) {
               justifyContent: "center",
               alignItems: "center",
             }}>
-            <Button variant="outlined">Upload Photo</Button>
-            <Button sx={{ marginTop: 1 }}>Remove</Button>
+            <TextField
+              ref={ref}
+              type="file"
+              accept="image/*"
+              onChange={handleImageChange}
+            />
+            {/* <Button variant="outlined">Upload Photo</Button> */}
+            <Button sx={{ marginTop: 1 }} onClick={handleRemoveImage}>
+              Remove
+            </Button>
           </Box>
         </Stack>
         <Stack spacing={2}>
@@ -132,9 +158,9 @@ const Personalize = function (props) {
               {price_estimations.map((price_estimate) => (
                 <FormControlLabel
                   key={price_estimate}
-                  value={price_estimate == "Yes"}
+                  value={price_estimate}
                   control={<Radio />}
-                  label={capitalCase(price_estimate)}
+                  label={price_estimate ? "Yes" : "No"}
                   sx={{ flex: 1 }}
                 />
               ))}
@@ -151,9 +177,9 @@ const Personalize = function (props) {
               {cleanings.map((cleaning) => (
                 <FormControlLabel
                   key={cleaning}
-                  value={cleaning == "Yes"}
+                  value={cleaning}
                   control={<Radio />}
-                  label={capitalCase(cleaning)}
+                  label={cleaning ? "Yes" : "No"}
                   sx={{ flex: 1 }}
                 />
               ))}
